requirements.txt
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.44.0
accelerate>=0.33.0
bitsandbytes>=0.43.0
pillow>=10.0.0
tqdm>=4.66.0
sentencepiece>=0.2.0
protobuf>=5.27.0
